requirements.txt
google-api-python-client
google-auth
google-auth-oauthlib
PyPDF2
requests
python-dotenv
google-cloud-vision
openai